// src/app/interfaces/app.dataCentros.interface.ts
export interface CentrosData {
    idCentroGestion:    number;
    idEstCentroGestion: number;
    descripcion:        string;
    codigo:             number;
    idMandante:         string;
    fecInicio:          string;
    fecTermino:         string;
    presupuestoContrato:  number;
    adicionalesAprobados: number;
    adicionalesEstimados: number;
    duenoObra:            string;
    rutEjecucion:         string;
}